+++ cinIO.py
@@ -2,12 +2,21 @@
 
 import json
 import os
+import yaml
 
 cachePath = os.path.join(os.path.dirname(__file__), str("cache\\"))
 configsPath = os.path.join(os.path.dirname(__file__), str("configs\\"))
 
 def loadConfig(fileName: str):
-    return json.load(open(os.path.join(configsPath + fileName)))
+    filePath = os.path.join(configsPath + fileName)
+    with open(filePath, 'r') as thisConfigFile:
+        if "json" in fileName:
+            return json.load(thisConfigFile)
+        elif "yaml" in fileName:
+            return yaml.safe_load(thisConfigFile)
+        else:
+            # todo: logging
+            print(f"invalid file format for config {filePath}")
 
 def loadCache(fileName: str):
     return json.load(open(os.path.join(cachePath + fileName)))
@@ -29,11 +38,13 @@ def joinWithGlobalVars(textsToJoin):
 
 
 
-token = loadConfig("token.json")["token"]
-config = loadConfig("config.json")
-strings = loadConfig("strings.json")
+token = loadConfig("token.yaml")["token"]
+config = loadConfig("config.yaml")
+strings = loadConfig("strings.yaml")
+minecraftConfig = loadConfig("minecraft.yaml")
+
 simpleResponses = loadConfig("simpleResponses.json")
-minecraftConfig = loadConfig("minecraft.json")
+
 reminders = loadCache("reminders.json")
 
 with open(os.path.join(os.path.dirname(__file__), str("assets\\conversation starters.txt")), "r") as file: